Migrate getPlaceholderState
