--- packages/addon/include/virtual_drive/Wrappers.h
@@ -1,10 +1,0 @@
-#pragma once
-
-#include <internal.h>
-#include <node_api.h>
-#include <virtual_drive.h>
-
-inline napi_value GetPlaceholderStateWrapper(napi_env env, napi_callback_info args)
-{
-    return NAPI_SAFE_WRAP(env, args, get_placeholder_state_wrapper);
-}
